+++ des/src/transitions.h
@@ -8,8 +8,8 @@ class Transition {
     public:
         Transition(std::string const& name, int to, NumericMatrix params);
         static Transition *create_transition(std::string const& dist, int to, NumericMatrix params);
-        float draw_event_time(int id) const;
-        virtual float draw(NumericMatrix::ConstRow row) const = 0;
+        double draw_event_time(int id) const;
+        virtual double draw(NumericMatrix::ConstRow row) const = 0;
 
         const std::string name;
         const int to;
@@ -19,37 +19,37 @@ class Transition {
 class WeibullTransition: public Transition {
     using Transition::Transition;
     public:
-        float draw(NumericMatrix::ConstRow) const;
+        double draw(NumericMatrix::ConstRow) const;
 };
 
 class LogNormalTransition: public Transition {
     using Transition::Transition;
     public:
-        float draw(NumericMatrix::ConstRow row) const;
+        double draw(NumericMatrix::ConstRow row) const;
 };
 
 class LogLogisticTransition: public Transition {
     using Transition::Transition;
     public:
-        float draw(NumericMatrix::ConstRow row) const;
+        double draw(NumericMatrix::ConstRow row) const;
 };
 
 class GammaTransition: public Transition {
     using Transition::Transition;
     public:
-        float draw(NumericMatrix::ConstRow row) const;
+        double draw(NumericMatrix::ConstRow row) const;
 };
 
 class GompertzTransition: public Transition {
     using Transition::Transition;
     public:
-        float draw(NumericMatrix::ConstRow row) const;
+        double draw(NumericMatrix::ConstRow row) const;
 };
 
 class ExpTransition: public Transition {
     using Transition::Transition;
     public:
-        float draw(NumericMatrix::ConstRow row) const;
+        double draw(NumericMatrix::ConstRow row) const;
 };
 
 #endif
